Allow specifying the outdir in bare compiler
Currently, we just write to ebin/ in the current working directory; this
implies that the source _should_ always be in the same subpath as the
ebin/ directory.

In the cases of utils like Mix, which store downloaded dependencies in
deps/ and artifacts in _build/, this may force more rebuild or less
caching that would be ideal.

This PR adds --outdir / -o as options to the bare compiler, which lets
the user specify another path, with a default kept to $CWD.