OTP-24 Builds passing
This does not get rid of all warnings.

This fixes the HiPE dependency to be conditional to whether dialyzer
actually requires it or not, and makes use of the .app.src.script
feature internally, at build time. This may cause issues for scripts
built prior to OTP-24 being used for OTP-24 versions as it will ask for
apps that aren't in place, but there's no good work-around for that.

This also bumps the erlware-commons version to one that eliminates some
warnings, and makes the bootstrap script honor the warnings_as_errors
value of dependencies to allow the build to proceed until parse_trans is
fixed upstream.