Fold rebar_uri calls into bootstrap script
Since the bootstrap script is executed before the compilation of
rebar_uri takes place, it can't rely on its functions. The functions are
used for the proxied calls and therefore need to be there before deps
required for rebar3 itself are fetched.

Folding it in is the only way to go while maintaining compatibility with
all versions, and we incur a mandatory warning for call paths that are
gonna be dynamically done but won't error out in practice.