{xref_ignores} now properly works on module, by taking into account function calls.
Summary: Before, the xref_ignores would properly ignores function calls on specified modules. That is,
even though {xref_ignores, [ModName,...]} would be in rebar3.config, those unassigned function calls
to and from the ModName module would be displayed.

This was due to a mismatch between the format of the ignore (a module name), and the format of the
warning ({MFA, MFA} or MFA) emmited by xref.