Ensure app vsn in app_info
This is a partial revert of commit 80671df

In commit 80671df, the app vsn setting after compile was
removed because it was believed that app vsn is set during app
discovery stage and hence no need to set again after compile.

This leads to an issue for elixir dependencies (by rebar_mix plugin)
which do not have version info during discovery.

In this commit, the `vsn` and `original_vsn` were provided
when validating modules as before commit 80671df, but only if
the fields were not set before.