add an option to select the application to build an script for
In the current state, in a multi-application project, the escriptize plugin
will only allow building a single application as an escript, and will force
developers to provide a "main" application with the escript_main_app
configuration variable.

This commit adds a command line option for the escriptize plugin which lets
the user select the application to build the escript for. The command line
option overrides escript_main_app if it is set.