Normalize Dialyzer PLT paths
this takes the current path formatting (which reuses compiler-specific
functions) and makes it contextual to knowing whether files exist or
not. The compiler path printing assumes all directory paths are
project-local, but PLTs only occasionally fit this description. On first
copies, they don't.

This currently yields output such as:

    ===> Building with 204 files in ../.cache/rebar3/rebar3_24.0.2_plt...
    ===> Copying ../.cache/rebar3/rebar3_24.0.2_plt to _build/default/rebar3_24.0.2_plt...
    ===> Checking 204 files in _build/default/rebar3_24.0.2_plt...
    ===> Adding 29 files to _build/default/rebar3_24.0.2_plt...
    ===> Doing success typing analysis...
    ===> Resolving files...
    ===> Analyzing 13 files with _build/default/rebar3_24.0.2_plt...

Whereas this patch instead allows the following formatting:

    ===> Dialyzer starting, this may take a while...
    ===> Updating plt...
    ===> Resolving files...
    ===> Updating base plt...
    ===> Resolving files...
    ===> Checking 204 files in /home/ferd/.cache/rebar3/rebar3_24.0_plt...
    ===> Copying /home/ferd/.cache/rebar3/rebar3_24.0_plt to /tmp/chk/_build/default/rebar3_24.0_plt...
    ===> Checking 204 files in _build/default/rebar3_24.0_plt...
    ===> Doing success typing analysis...
    ===> Resolving files...
    ===> Analyzing 2 files with _build/default/rebar3_24.0_plt...

That `Copying <from> to <dest>` message is always going to have the
`<dest>` value be absolute just because I can check whether a relative
path exists before normalizing on the absolute path, which is the
simplest way to make this work without major rework of output functions
to attach them all to the calling context to know when a path may or may
not be absolute.

Fixes https://github.com/erlang/rebar3/issues/2577