Make escript_incl_extra work with _checkouts
When developing `relx` and testing changes with `rebar3`, it is convenient to make `relx` a checkout dependency.

However, the current implementation of `escript_incl_extra` expects the output directory of `relx` to be under `_build/<profile>/lib/`, which is not the case if `relx` is a checkout dependency.

Instead of making a breaking change to `escript_incl_extra`, introduce a new config parameter instead.