Build compiler order DAG without topological sort
As described in https://github.com/erlang/rebar3/issues/2563,
part of the previous approach's weakness is that we need to
seed the compile order with the dependency order for some
invisible compile-time dependencies (runtime ones turned
to build-time ones via calls to parse-transforms).

Unfortunately, topological sorting is insufficient due to how it
flattens apps. Assume the DAG contains:

    a -> b
    c -> d

then a valid topological sort would be `[a,c,b,d]`. However, if we
introduce the hidden dep (and decide that `c` calls `b`in a
parse_transform) that the DAG doesn't contain, then the sorting order
is broken.

However, if we initially carried the `[x,b,a,c,f,e,d]` dep list,
we can use a run over it to build `[x,a,b,c,f,e,d]` final compile
order.