Remove duplicated message when trying to upgrade a checkout dep
When running `rebar3 upgrade` on a project that has checkout dependencies,
a message of `"App $NAME is a checkout dependency and cannot be locked."`
will be printed by the `lock` provider.

As this provider is called twice (first to check the locked dependencies,
and then from the `upgrade` provider after updating the dependencies),
the message will show duplicated. To avoid this, this commit marks
all checkout dependencies as regular ones, after checking whether it makes
sense to do so, before calling the `lock` provider from the `upgrade`
provider.