Fix global cache config overriding
There was some confusion in the code:

- `rebar_dir:global_config()` returns the default path
- `rebar_dir:global_config(State)` returns the configured path based on
  the ENV overrides, and defaults to `global_config()` if nothing is
  given

Since no State term existed at this point in time, we just called out
directly to the argument-free version of the call, but this ignored all
overrides. Instead, we call it with an initialized `rebar_state:new()` to
properly handle configuration from users, and then use the actual
configured state (if any).