Local plugins can't declare project plugins
This adds a test confirming the implemented behaviour, which states that
a local plugin that declares a project_plugin won't see it fetched from
the root project. This would generally be a thing ignored by a
dependency and it makes no sense to have it declared in the local plugin
and also impact the top project.