scope local install VSN variable
See https://github.com/erlang/rebar3/issues/2751

By default we use the `VSN` variable and let it be overridden. But that variable is not prefixed to anything rebar3 so unrelated variables from random scripts are possible to break by doing this. 

I believe `VSN` is not documented anywhere, so we should be safe in changing it  right away.

CC @tsloughter