Add test case for copying ct `_SUITE_data` in app/test
The tests checks that the files and the symbolic files are copied
correctly to the _build folder.

In the case of symbolic links, the copy needs to dereference the
content. If not, the symbolic link can point to relative path that
doesn't exist in _build folder.

These tests try to prove that the feature added in [1] doesn't work
for all cases. When the `_SUITE_data` is inside an app test directory,
the content are copied without referencing.

[1] https://github.com/erlang/rebar3/pull/2731