Attempt to support legible compiler error messages
This is rather straightforward as an implementation: given any error
message that contains column information, read the source file and find
that line. Display it, and indent a cursor underneath it that points to
the column.

Then we can just show the same error message as before.

In practice, given the module:

    -module(rebar_fake).

    -export([diagnostic/1]).

    diagnostic(A) ->
        X = add(5 / 0),
        {X,X}.

    add(X) -> X.

    add(X, Y) -> X + Y.

Calling rebar3 compile yields:

    ...
    ===> Compiling rebar
    ===> Compiling apps/rebar/src/rebar_fake.erl failed
    apps/rebar/src/rebar_fake.erl:5:12:
      diagnostic(A) ->
                 ^-- variable 'A' is unused
    apps/rebar/src/rebar_fake.erl:6:15:
          X = add(5 / 0),
                    ^-- evaluation of operator '/'/2 will fail with a 'badarith' exception
    apps/rebar/src/rebar_fake.erl:11:1:
      add(X, Y) -> X + Y.
      ^-- function add/2 is unused

This is friendlier than what we had before for sure. Adding color could
be an extension to this if we find the format useful.

The rebar compiler output richness is conditional and configurable, and we
default to the minimal normal one to  make sure it won't break all sorts of tooling.

Rich errors also happen to support to xrl and yrl compilers