Maintain cth_log_redirect default in CT runs
By default, common test should be redirecting log output via the
cth_log_redirect hook. However, whenever we override the values to
enable the stuff in cth_readable, we take that log redirection away.

This restores it so it is the same, so long as the user does not specify
their own log values in it, which keeps the override/removal as-is.