Hotfix erlware_commons config script
In switching to vendoring, we implicitly broke the configuration-based
version checks that rely on the app vsn of rebar3. this app vsn isn't
configured yet during the initial app discovery phase and the "git"
value is seen as invalid, causing a bootstrap crash.

We didn't see it because since Rebar3 3.18.0, we had not returned to the
"git" version for things built from main, which hid the weird behavior.

It will legitimately happen absolutely nowhere but in this repo, so
until it is modified elsewhere I'm hotpatching the dep to fix nightly
builds.