Defer hostname check to OTP when supported
While investigating the work required to support
https://github.com/erlang/rebar3/pull/2803, I found out that the code
was already in place.

However, despite the code being there, we still passed the old
`ssl_verify_hostname:verify_fun/3` function of pre-21.0 on top of it,
which I supposed ignored the check.

So this change reworks the flow such that we fall back to the legacy
check only if it isn't supported by the OTP library at this point.
Getting this going would require someone to build a new release on an
Erlang copy older than OTP-21 (which is no longer supported) which is
unlikely.

This follows guidelines from
https://erlef.github.io/security-wg/secure_coding_and_deployment_hardening/ssl