Dedupe compiler DAG edge insertion for artifacts
Issue described at
https://erlangforums.com/t/rebar3-is-taking-longer-to-compile-projects/3061/13

It appears that edge insertion with a label involved messes with the bag
table and continuously inflates the size of a DAG after many runs. Since
changes are propagated through a graph, frequent changes to
often-included header files or parse transforms risk multiplying edge
counts.

This little patch ensures that before inserting an artifact edge, we
first look whether it already exists; if so we don't add it any further,
but keep all other semantics the same.