introduce rebar_semver module
Instead of calling ec_semver directly, all version-related
calls go through this module; This makes it easy
to swap later when hex_core exposes an appropriate
replacement in the future.

It was not possible to use `verl` or other libraries
here, since rebars/erlang_commons parsing was
different in a number of edge cases:

- rebar allows version bounds with 2 components,
  like `< 1.0` or `== 1.2`.
- rebar allows approximate bounds with more than 3
  components, like `~> 1.5.6-rc.0`

To continue to support these, the new module also
uses ec_semver under the hood, but extends the custom
constraint parser to also support `and` and `or`
constraints used by Elixir or Gleam.

Constraints are modelled as a match function without
a pure data representation. This has to be assumed
to be an implementation detail that might change in
the future when a need arises.